Extract basic types, constraints, and solve items from statements
