+++ src/primitive_literals.rs
@@ -285,7 +285,7 @@ fn test_float() {
     assert_eq!(float_literal::<VerboseError<&str>>("1.0,"), Ok((",", 1.0)));
 
     assert_eq!(float_literal::<VerboseError<&str>>("0.000000000000000000000000000000007609999999000000000000000000000000760999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999900000000000000000000000764DD4DDDDDDDDD%
-    "), Err(crate::Err::Error(VerboseError { errors: vec![("0.000000000000000000000000000000007609999999000000000000000000000000760999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999900000000000000000000000764DD4DDDDDDDDD% \n    ", nom::error::VerboseErrorKind::Nom(nom::error::ErrorKind::MapRes))] })));
+    "), Err(crate::Err::Error(VerboseError { errors: vec![("0.000000000000000000000000000000007609999999000000000000000000000000760999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999900000000000000000000000764DD4DDDDDDDDD%\n    ", nom::error::VerboseErrorKind::Nom(nom::error::ErrorKind::MapRes))] })));
 }
 
 pub fn float_literal<'a, E: ParseError<&'a str>>(input: &'a str) -> IResult<&'a str, f64, E>
